# openapi.yaml
openapi: 3.0.0
info:
  title: 陌生人聊天App关注服务
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: 生产环境
  - url: https://staging-api.example.com
    description: 预发布环境
  - url: http://localhost:8089
    description: 本地开发环境
paths:
  /api/v1/follow/user:
    post:
      summary: 关注用户
      description: 用户通过此接口关注其他用户
      operationId: followUser
      security:
        - jwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - targetUserId
              properties:
                targetUserId:
                  type: string
                  format: uuid
                  description: 目标用户ID，唯一标识被关注的用户
                  example: "123e4567-e89b-12d3-a456-426614174001"
                  minLength: 36
                  maxLength: 36
      responses:
        '200':
          description: 关注成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 响应状态
                    example: "success"
                    enum:
                      - "success"
                      - "fail"
                  message:
                    type: string
                    description: 响应消息
                    example: "关注成功"
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "请求参数错误"
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "服务器内部错误，请稍后再试"
    delete:
      summary: 取消关注用户
      description: 取消当前用户对目标用户的关注
      operationId: unfollowUser
      security:
        - jwtAuth: [ ]
      parameters:
        - in: query
          name: targetUserId
          schema:
            type: string
            format: uuid
            description: 目标用户ID，唯一标识目标用户的字符串
            required: true
            example: "123e4567-e89b-12d3-a456-426614174001"
            minLength: 36
            maxLength: 36
      responses:
        '200':
          description: 取消关注成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 响应状态
                    example: "success"
                    enum: [ "success", "fail" ]
                  message:
                    type: string
                    description: 响应消息
                    example: "取消关注成功"
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "参数缺失或格式错误"
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "服务器内部错误，请稍后再试"
  /api/v1/follow/my-follows:
    get:
      summary: 获取我的关注列表
      description: 获取当前登录用户的关注列表
      security:
        - jwtAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            description: 返回的关注数量限制
            example: 10
            default: 10
          required: false
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            description: 偏移量，用于分页
            example: 0
            default: 0
          required: false
      responses:
        '200':
          description: 成功获取关注列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  follows:
                    type: array
                    items:
                      type: object
                      properties:
                        targetUser:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                              description: 被关注用户的ID
                              example: "123e4567-e89b-12d3-a456-426614174000"
                            avatar:
                              type: string
                              format: uri
                              description: 被关注用户的头像URL
                              example: "https://example.com/avatars/user123.jpg"
                            username:
                              type: string
                              minLength: 1
                              maxLength: 30
                              description: 被关注用户的用户名（若未设置则显示手机号后四位）
                              example: "john_doe"
                        latestPostContent:
                          type: string
                          format: string
                          description: 被关注用户最新发表的帖子的文字内容
                          example: "这是一个新的帖子内容"
                          minLength: 1
                          maxLength: 200
                        timestamp:
                          type: string
                          format: date-time
                          example: '2023-10-01T12:00:00Z'
                          description: 用户关注该对象的时间戳
                  totalCount:
                    type: integer
                    description: 总关注数
                    example: 100
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "参数缺失或格式错误"
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "服务器内部错误，请稍后再试"
  /api/v1/follow/my-fans:
    get:
      summary: 获取我的粉丝列表
      description: 获取当前登录用户的粉丝列表
      security:
        - jwtAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            description: 返回的粉丝数量限制
            example: 10
            default: 10
          required: false
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            description: 偏移量，用于分页
            example: 0
            default: 0
          required: false
      responses:
        '200':
          description: 成功获取粉丝列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  fans:
                    type: array
                    items:
                      type: object
                      properties:
                        targetUser:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                              description: 粉丝用户的ID
                              example: "123e4567-e89b-12d3-a456-426614174000"
                            avatar:
                              type: string
                              format: uri
                              description: 粉丝用户的头像URL
                              example: "https://example.com/avatars/user123.jpg"
                            username:
                              type: string
                              minLength: 1
                              maxLength: 30
                              description: 粉丝用户的用户名（若未设置则显示手机号后四位）
                              example: "john_doe"
                        latestPostContent:
                          type: string
                          format: string
                          description: 粉丝用户最新发表的帖子的文字内容
                          example: "这是一个新的帖子内容"
                          minLength: 1
                          maxLength: 200
                        timestamp:
                          type: string
                          format: date-time
                          example: '2023-10-01T12:00:00Z'
                          description: 关注时间
                  totalCount:
                    type: integer
                    description: 总粉丝数
                    example: 100
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "参数缺失或格式错误"
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "服务器内部错误，请稍后再试"
  /api/v1/follow/mutual:
    get:
      summary: 获取互相关注列表
      description: 获取当前登录用户的互相关注列表
      security:
        - jwtAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            description: 返回的互相关注数量限制
            example: 10
            default: 10
          required: false
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            description: 偏移量，用于分页
            example: 0
            default: 0
          required: false
      responses:
        '200':
          description: 成功获取互相关注列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  mutualFollows:
                    type: array
                    items:
                      type: object
                      properties:
                        targetUser:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                              description: 互相关注的用户ID
                              example: "123e4567-e89b-12d3-a456-426614174000"
                            avatar:
                              type: string
                              format: uri
                              description: 互相关注的头像URL
                              example: "https://example.com/avatars/user123.jpg"
                            username:
                              type: string
                              minLength: 1
                              maxLength: 30
                              description: 互相关注的用户名（若未设置则显示手机号后四位）
                              example: "john_doe"
                        latestPostContent:
                          type: string
                          format: string
                          description: 互相关注的用户最新发表的帖子的文字内容
                          example: "这是一个新的帖子内容"
                          minLength: 1
                          maxLength: 200
                        timestamp:
                          type: string
                          format: date-time
                          example: '2023-10-01T12:00:00Z'
                          description: 互相关注的时间戳
                  totalCount:
                    type: integer
                    description: 总互相关注数
                    example: 100
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "参数缺失或格式错误"
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "服务器内部错误，请稍后再试"
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      name: Authorization
      in: header
      description: 使用JWT进行认证，将JWT凭证放在请求头中